Pushing all of my updates and explorations from this week.
The lorenz96 folder has smaller studies focusing on attempting to
reimplement the dynamics and the TorchODE integration + Neural ODE
training. The main src/mace folder has been pared down to just the
necessary files for the lorenz 96 implemenation, but for some reason the
dynamics are still not being learned correctly.